==> Experiment-2_Analysis/01_model-selection.R <==
#################################
#     model_selection           #
#################################
# This script does:
# finds random effects structure for (g)lmm
# and saves final models in models/
#
# Author: Aylin Kallmayer
# Year: 2022

# load functions
source("src/01_mixed-models_functions.R")

# run to preprocess new data
# source("00_preprocessing.R")

if (!require("pacman")) install.packages("pacman")
p_load("dplyr","lmerTest")

#----ACCURACY----
#----badsrem outrem data----
data <- prepare_for_model(data_type="badsrem_outrem")

# Find RE structure
summary(acc.model <- glmer(correct ~ rotation*consistency + # fixed effects
                             (1 + rotation + consistency | sub) +
                             (1 + rotation + consistency | imgID), # random effects
                           family = binomial(),
                           control=glmerControl(optimizer="bobyqa"),
                           data = data))
summary(rePCA(acc.model)) #deprecate
summary(acc.model.2 <- glmer(correct ~ rotation*consistency + # fixed effects
                             (1 + consistency | sub) +
                             (1 + rotation + consistency | imgID), # random effects
                           family = binomial(),
                           control=glmerControl(optimizer="bobyqa"),
                           data = data))
anova(acc.model, acc.model.2) #not sig worse
summary(rePCA(acc.model.2)) #deprecate
summary(acc.model.3 <- glmer(correct ~ rotation*consistency + # fixed effects
                               (1 + consistency | sub) +
                               (1 + rotation | imgID), # random effects
                             family = binomial(),
                             control=glmerControl(optimizer="bobyqa"),
                             data = data))
anova(acc.model.2,acc.model.3)
summary(rePCA(acc.model.3)) #deprecate
summary(acc.model.4 <- glmer(correct ~ rotation*consistency + # fixed effects
                               (1 | sub) +
                               (1 + rotation | imgID), # random effects
                             family = binomial(),
                             control=glmerControl(optimizer="bobyqa"),
                             data = data))
anova(acc.model.3,acc.model.4)
summary(rePCA(acc.model.4)) #not deprecate

# final model
summary(acc.model.4)
saveRDS(acc.model.4, "models/acc-model_processed_badsrem_outrem.rds")

#----RT----
#----badsrem Data----
data <- prepare_for_model(data_type="badsrem", RT = TRUE)

#----Find RE structure----
summary(rt.model <- lmer(log_RT ~ rotation*consistency + # fixed effects
                           (1 + rotation + consistency | sub) +
                           (1 + rotation + consistency | imgID), # random effects
                         data = data))
summary(rePCA(rt.model)) # deprecate
summary(rt.model.2 <- lmer(log_RT ~ rotation*consistency + # fixed effects
                           (1 + rotation | sub) +
                           (1 + rotation + consistency | imgID), # random effects
                         data = data))
anova(rt.model, rt.model.2)
summary(rePCA(rt.model.2)) # not deprecate

# final model
summary(rt.model.2)
saveRDS(rt.model.2, "models/rt-model_processed_badsrem.rds")
