==> Experiment-2_Analysis/02_vis-and-posthoc.R <==
#################################
#        plots & post-hoc       #
#################################
# This script does:
# 1) post-hoc tests
# 2) accuracy and rt plots
# 3) model adjusted accuracy and rt plots
#
# Author: Aylin Kallmayer
# Year: 2022

if (!require("pacman")) install.packages("pacman")
p_load("dplyr","ggplot2","emmeans","remef","ggeffects","Rmisc")

# functions
source("src/01_mixed-models_functions.R")

#----Preprocess----
# run to preprocess new data
# source("00_preprocessing.R")

# update ggplot theme
global_theme <- theme_classic()+
  theme(axis.title=element_text(size=12),
                      axis.text=element_text(size=10),
                      legend.position = "None")

#----ACCURACY----
# with out badsrem data
data <- prepare_for_model(data_type="badsrem_outrem")
acc.model <- readRDS("models/acc-model_processed_badsrem_outrem.rds")
summary(acc.model)

#----Post-hoc tests----
(emm_1 <- emmeans(acc.model, pairwise ~ rotation | consistency))
(emm_2 <- emmeans(acc.model, pairwise ~ consistency | rotation))
(emm_2 <- emmeans(acc.model, pairwise ~ consistency * rotation))

(IC_st <- contrast(emm_1[[1]], interaction = "pairwise", by = NULL))

# -> the viewpoint effect is significantly stronger in the inconsistent than the consistent condition

#----PLot adjusted accuracy---- 
(a <- ggpredict(acc.model, terms=c("consistency","rotation")) %>% ggplot(aes(x, predicted, color=group))+
   geom_point(position=position_dodge(.5))+
   geom_errorbar(aes(ymax=conf.high, ymin=conf.low), width=.5, position = position_dodge(.5))+
   scale_color_manual(values = c("#56b4e9", "#cc79a7"))+
   xlab("Consistency")+
   ylab("Adjusted Accuracy Predictions")+
   ggtitle("")+
   global_theme)
tiff("plots/acc_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
a# insert ggplot code
dev.off()

#----PLot difference score---- 
data_diff <- data %>% group_by(sub, consistency, rotation) %>%
  dplyr::summarise(mean_acc = mean(correct),
            .groups="keep") %>%
  ungroup() %>%
  group_by(sub, consistency) %>%
  dplyr::mutate(diff_rot = mean_acc[rotation=="canon"]-mean_acc[rotation=="noncanon"]) %>%
  dplyr::summarise(diff = mean(diff_rot),
            .groups="keep")
table <- summarySEwithin(data = data_diff, measurevar = "diff", withinvars = "consistency")

(x <- ggplot(data_diff, aes(x=consistency, y=diff))+
  geom_violin(aes(fill=consistency), alpha=.4)+
    geom_boxplot(width=.3)+
    geom_point(alpha=.1)+
    geom_line(aes(group=sub, color=as.factor(sub)), alpha=.5)+
    #stat_summary(fun = "mean", colour = "black", size = 2, geom = "point")+
    #geom_point(data=table, size=2)+
    #geom_errorbar(data=table, aes(ymin=diff-se, ymax=diff+se), width=.1)+
    #geom_line(data=table, aes(group=1))+
    xlab("Scene Consistency")+
    ylab("Difference Score \n (canonical vs. non-canonical)")+
    scale_fill_manual(values = c("#56b4e9", "#cc79a7"))+
    global_theme)
tiff("plots/acc_diffplot.png", units="cm", width=10, height=8, res=300)
x# insert ggplot code
dev.off()

#----RT----
# with bads removed
data <- prepare_for_model(data_type="badsrem", RT=TRUE)
rt.model <- readRDS("models/rt-model_processed_badsrem.rds")
summary(rt.model)

#----PLot RT---- 
(b <- ggpredict(rt.model, terms=c("rotation")) %>% ggplot(aes(x, predicted, color=x)) +
  geom_point()+
  geom_errorbar(aes(ymax=conf.high, ymin=conf.low), width=.5)+
  scale_color_manual(values = c("#56b4e9", "#cc79a7"))+
  xlab("Viewpoint")+
  ylab("Adjusted log RT Predictions")+
  ggtitle("")+
  global_theme)
tiff("plots/rt_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
b# insert ggplot code
dev.off()

# adj violin plot
data$RT_adj <- keepef(rt.model, fix=c("rotation1","consistency1"), keep.intercept = FALSE)
table <- aggregate(data=data, RT_adj ~ sub+rotation+consistency, mean) %>%
  arrange(consistency, sub)%>%
  mutate(group=rep(1:64, each=2))
(c <- ggplot(table, aes(x=consistency, y=RT_adj, fill=rotation))+
  geom_violin(alpha=.4)+
    geom_boxplot(position = position_dodge(.9), width=.3, fill="white", aes(color=rotation))+
    scale_color_manual(values=c("black","black"))+
    xlab("Scene Consistency")+
    ylab("Adjusted Response Time")+
    geom_point(alpha=.1,aes(color=as.factor(rotation)), position=position_dodge(.9), 
               color="black")+
    scale_fill_manual(values = c("grey", "black"))+
    global_theme)
png("plots/rt_adjusted-badsrem_outrem.png", units="cm", width=10, height=8, res=300)
c# insert ggplot code
dev.off()
